[dashboard] Don't show User Notifications and Feedback on Dedicated
Tool: gitpod/catfood.gitpod.cloud
