+++ components/dashboard/src/Insights.tsx
@@ -26,7 +26,7 @@ import { useTemporaryState } from "./hooks/use-temporary-value";
 import { DownloadIcon } from "lucide-react";
 import { Button } from "@podkit/buttons/Button";
 import { DropdownMenu, DropdownMenuTrigger, DropdownMenuContent, DropdownMenuItem } from "@podkit/dropdown/DropDown";
-import { useInstallationConfiguration } from "./data/installation/default-workspace-image-query";
+import { useInstallationConfiguration } from "./data/installation/installation-config-query";
 
 export const Insights = () => {
     const toDate = useMemo(() => Timestamp.fromDate(new Date()), []);
